Ensure an error is emitted when parsing a stmt after a last stmt in the top level block (#302)
There is code that "merges" two top level blocks when performing initial
parsing. This code does not account for the fact that a stmt parsed
after a last stmt is invalid syntax. We extend it to ensure an error is
reported in this case.

Note that we do not resolve the problem where the stmts after the last
stmt "teleport" in front of the last stmt in the AST. That is more
difficult to sort without restructuring the AST. However this was
originally parsed as invalid syntax so we now at least emit an error and
do not guarantee the state of an invalid syntax tree.

Partially resolves #298