Support parsing Luau function attributes (#337)
Supports the parsing of Luau function attributes, such as `@native` or
`@checked`. Does not perform any validation on the name of attributes,
unlike Luau.

Unfortunately, due to the existing `Attribute` struct for Lua 5.4
`<const>` attributes, we have to name the struct `LuauAttribute`.

**[BREAKING CHANGE]**: Because attributes can be attached to anonymous
functions, we create a new `AnonymousFunction` struct to handle this.
This replaces the old `Box<(TokenReference, FunctionBody)>`.

Closes #336